requirements.txt
Flask==0.10.1
Jinja2==2.7.1
MarkupSafe==0.18
Werkzeug==0.9.4
argparse==1.2.1
google-api-python-client==1.2
gunicorn==18.0
httplib2==0.8
itsdangerous==0.23
python-gcm==0.1.4
requests==2.0.1
sleekxmpp==1.1.11
wsgiref==0.1.2
